+++ src/include/knowledge/tagdict.h
@@ -24,32 +24,32 @@ class CTagDict {
 
 protected:
 
-   CHashMap<CWord, unsigned long> m_mapItems;
+   CHashMap<CWord, unsigned long long> m_mapItems;
    unsigned long m_nTotalNumberOfTags;
 
 public:
 
    CTagDict(unsigned long nTotalNumberOfTags) { 
       m_nTotalNumberOfTags = nTotalNumberOfTags ;
-      assert( nTotalNumberOfTags < sizeof(unsigned long) * 8 ); 
+      assert( nTotalNumberOfTags < (sizeof(unsigned long long)*8) );   
    }
 
    virtual ~CTagDict() { }
 
 public:
 
-   unsigned long lookup(const CWord &word) const {
+   unsigned long long lookup(const CWord &word) const {
       return m_mapItems.find(word, 0);
    }
 
    // whether the word and tag pair is in the dictionary
    bool lookup(const CWord &word, const CTag &tag) const {
-      return m_mapItems.find(word, 0) & (static_cast<unsigned long int>(1)<<tag.code()) ;
+      return m_mapItems.find(word, 0) & (static_cast<unsigned long long>(1)<<tag.code()) ;
    }
 
    // add a word and tag pair into entries
    void add(const CWord &word, const CTag &tag) {
-      m_mapItems[word] |= (static_cast<unsigned long>(1)<<tag.code());
+      m_mapItems[word] |= (static_cast<unsigned long long>(1)<<tag.code());
    }
 
 public:
@@ -57,7 +57,7 @@ public:
    // load tag dictionary from a stream
    void load(istream &is) {
       string s, word, tag ;
-      unsigned long int i ;
+      unsigned long long i ;
       getline(is, s);
       while( is && !(s.empty()) ) {
          istringstream iss(s);
@@ -66,7 +66,7 @@ public:
 //         if ( i==0 ) {
 //            throw("unrecognized tag in the dictionary file: " + tag + ".");
 //         }
-         m_mapItems[word] |= (static_cast<unsigned long int>(1)<<i) ;
+         m_mapItems[word] |= (static_cast<unsigned long long>(1)<<i) ;
          getline(is, s);
       }
    }
@@ -83,13 +83,13 @@ public:
 
    // save tag dictionary to a stream
    void save(ostream &os) {
-      CHashMap<CWord, unsigned long>::iterator it;
-      unsigned long i, j ; 
+      CHashMap<CWord, unsigned long long>::iterator it;
+      unsigned long long i, j ; 
       it = m_mapItems.begin() ;
       while (it != m_mapItems.end()) {
          j = it.second();
          for (i = 0; i < m_nTotalNumberOfTags; ++i) {
-            if ( j & (static_cast<unsigned long int>(1)<<i) ) 
+            if ( j & (static_cast<unsigned long long>(1)<<i) ) 
                os << it.first().str() << "\t" << CTag(i).str() << endl ;
          }
          ++it ;
